+++ packages/state-manager/src/sagas/publicChannels/createChannel/createChannel.saga.ts
@@ -1,12 +1,13 @@
-import { type publicChannelsActions } from '../publicChannels.slice'
+import { publicChannelsActions } from '../publicChannels.slice'
+import { messagesActions } from '../../messages/messages.slice'
 import { type PayloadAction } from '@reduxjs/toolkit'
-
-import { apply } from 'typed-redux-saga'
+import { apply, put } from 'typed-redux-saga'
 
 import { type Socket, applyEmitParams } from '../../../types'
-
 import logger from '../../../utils/logger'
-import { SocketActionTypes } from '@quiet/types'
+
+import { SocketActionTypes, type CreateChannelResponse } from '@quiet/types'
+
 const log = logger('publicChannels')
 
 export function* createChannelSaga(
@@ -14,11 +15,27 @@ export function* createChannelSaga(
   action: PayloadAction<ReturnType<typeof publicChannelsActions.createChannel>['payload']>
 ): Generator {
   log(`Creating channel ${action.payload.channel.name}`)
-  yield* apply(
+
+  const response = yield* apply(
     socket,
-    socket.emit,
+    socket.emitWithAck,
     applyEmitParams(SocketActionTypes.CREATE_CHANNEL, {
       channel: action.payload.channel,
     })
   )
+
+  if (response) {
+    yield* put(
+      messagesActions.addPublicChannelsMessagesBase({
+        channelId: response.channel.id,
+      })
+    )
+    yield* put(publicChannelsActions.addChannel(response))
+    yield* put(
+      publicChannelsActions.sendInitialChannelMessage({
+        channelName: response.channel.name,
+        channelId: response.channel.id,
+      })
+    )
+  }
 }