+++ CHANGELOG.md
@@ -1,8 +1,30 @@
+[unreleased]
+
+# Fixes:
+
+* Remove duplicate introduction messages once again
+
+# Refactorings:
+
+* Rename message retrieval events/sagas and use ack feature
+* Rename LOAD_MESSAGES event and incomingMessages reducer
+* Use socket.io ack for CREATE_CHANNEL event
+* Introduce socket.io acknowledgements
+
+# Fixes:
+
+* Prevent channel creation with names that start with special character, then a hyphen
+
 [2.1.1]
 
 # Fixes:
 
+* Make sure address of the inviting peer is in the invitation link
 * Opening the mobile app with joining links has been corrected.
+
+# Refactorings:
+
+* Remove unused backend events and state-manager event types
 
 [2.1.0]